+++ src/components/pages/passwordForgot/PasswordForgot.jsx
@@ -1,0 +1,5 @@
+import React from "react"
+
+export default function PasswordForgot() {
+  return <div>PasswordForgot</div>
+}
